Deploy most files to GitHub pages.

+++ .github/deploy.sh
@@ -3,16 +3,19 @@
 set -ue
 
 cd "$(dirname "$0")"
-P=../pages
+P="$(mktemp -d)"
 
-cp deploy_key ~/.ssh/id_rsa
-chmod 600 ~/.ssh/id_rsa
+if [ -e deploy_key ]
+then
+	cp deploy_key ~/.ssh/id_rsa
+	chmod 600 ~/.ssh/id_rsa
+fi
 
 [ -d $P ] && rm -Rfv $P
 git clone --depth=10 --branch=gh-pages git@github.com:TobiX/jenkins-neo2-theme.git $P
 
-rsync -r --del --verbose ../dist/* $P/dist/
-cp ../README.md $P/index.md
+rsync -r --del --verbose --exclude node_modules --exclude less ../* $P/
+mv $P/README.md $P/index.md
 
 cd $P
 
@@ -22,4 +25,4 @@ git config push.default simple
 
 git add -A .
 git commit -a -m "Update website from commit $TRAVIS_COMMIT"
-git push origin HEAD:gh-pages
+[ -e deploy_key ] && git push origin HEAD:gh-pages
